Dropped constraints on "parent_id" and "blog_id" on "comments" table
Because of the polymorphic relations between "likes" & "blogs", and "likes" & "comments", cascadeOnDelete wasn't deleting the related rows on "likes" table, as there can be no constraints in polymorphic relations. To troubleshoot this, manual deletion needs to be implemented. This is the first step of the process to make comments independent so they can be manually deleted.