Re-added foreign key on constraints to the "comments" table
Because of my lack of knowledge I removed the foreign key constraints on "blog_id" and "parent_id" from the "comments" table. Re-added the foreign key to maintain data integrity. This time "ON DELETE" is set to no action so the parent record can't be deleted if all of it's children aren't already deleted. This flow will be implemented on the application in the next commit.