Changed triggering events from "deleted" to "deleting", changed filenames accordingly
Now BlogDeleting(previously BlogDeleted) & CommentDeleting(previously CommentDeleted) events will be triggered before parent record (blog/comment) is deleted. This will make sure the changes on the previous commit's migration is seamlessly implemented as the actual deletion chain will start from the last descendant and work it's way up the ancestry. In other words, replies to parent comment's replies will be deleted first, then replies to parent comments will be deleted, after that the parent comments will be deleted and finally the blog will be deleted. This is fully compatible with the "no action" behavior "ON DELETE" of the foreign keys which was set up in the migration in the last commit.

filenames and class names changed to reflect triggering of the events
-BlogDeleted to BlogDeleting
-CommentDeleted to CommentDeleting